Update swift keystoneauth options
In our previous releases, we were not using keystoneauth to validate
user role, so anybody with a valid token was allowed to use Swift. Now
we'll allow _member_ in, to match that functionality.

Also update the authtoken config to remove the old format of urls.

Remove the quotes on the cert file, as this seems to trip up requests.

(cherry picked from commit c8b8197a8ce29ee5c7ac99d64608cd4dcaf9bc76)