Support ip allocation pool definition for subnets
In some environments we cannot use the default IP allocation of the
entire subnet, and instead need to define a start and end to the pool.
When no start/end have been defined, do not pass along the options to
the module.

THe subnet creation task is duplicated here becuase Ansible 1.7 does not
have the "omit" filter.