Install kernel debug symbols for fast crashdumps
The dbgsym ddeb is large and on a slow mirror, and this only needs
to be done upon kernel upgrade. Without debug symbols, kdump is forced
to use a much slower page filtering method with makedumpfile.

Also increases reserved crashkernel memory to 512M--though this may not
be sufficient for systems with >128GB RAM, so this likely needs to be
dynamic. If makedumpfile is oomkilled, kdump will simply `cp
/proc/vmcore`, taking even longer.