Trigger service restarts on openstack code change
If we change the package or the git repo we want to restart the
services. Currently this also depends on 'upgrade' being set to True.

For neutron, we have to manually kill neutron-ns-metadata-proxy as it
doesn't get restarted when the l3 agent gets restarted. However we do
want it to be running from the new code.

(cherry picked from commit 42e49bdac8dbe489d230510f4dbb71f94ac072f9)