Make swift great again
Swift was previously using swift-init to manage all of the swift
services. Well it turns out that this leads to some confusion when using
upstart.

To address this, call all of the native binaries in the upstart_services
and reorder when we build the rings. This moves the start/restart
control back into Ansible's hands for now.

Some other small swift updates to make Vagrant great again.