Add mysql to haproxy (#2277)
This is a MVP for mysql behind haproxy, so that openstack services can
proxy through and survive a single node's mysql restart. Before this,
all nodes would point at the internal floating IP and connect to the
mysql process there directly. If mysql was restarted, all connections
would die. With a proxy, they can retry / reconnect to another member of
the cluster.

The backend checking isn't completely robust, there is a FIXME in place,
but this gets us started and testing.

Only OpenStack services (things that consume endpoints.db) are set to
use the proxy.

Change-Id: Ibf066ec4f515ab585a85a8061a16db12b8397cb2