metrics scheme was getting set with .'s on rhel
this screwed up our dashboards etc,  graphite treats .'s as seperators
and effort is made to ensure host metrics ship without .'s in the scheme.

unfortunately under rhel the FQDN was being used,  this fixes it to use the
newer `hostname_short` variable and if that doesn't exist defaults back to
using `ansible_nodename`.