Set selinux labels for services and enable enforcing mode
Add the selinux labels for swift and enable enforcing mode

modify the selinux label context for swift /srv/node dir

reverting the changes made for selinux

label the swift /srv/node dir after it is create

The openstack-selinux package provides the SELinux targeted policy
for swift. This policy has the correct context for /srv/node/*
path. The label does not get applied as the mount point is created
later. The change here runs the restorecon tool to apply the
context already specified in the swift targeted policy

Adding default variables for enablement of selinux

port sefcontext module and use it

Port the sefcontext module to ursula
Change the tasks to use it
Address the comments from review about output of restorecon commands

set selinux to permissive mode for centos