Increase buffer size for Keystone uwsgi
This increases the buffer size for Keystone uwsgi. The default for uwsgi
is 4096. The new size will be 114688 bytes, which will match the default
size limit setting for the oslo 'sizelimit' middleware which exists in
the pipline. This will fix some issues where we get dropped federated
requests to Keystone because of SAML assertion sizes.