set pool_names completed flag at the end of pool_name.yml
Ansible passes the parent condition to every sub task in included yml
file. In following example, `ssd_pool is undefined` is passed to every
in pool_names.yml as condition of every task in pool_names.yml.
So we have to defined the flag at the end of the yml file, to make sure
all tasks in the yml file is executed.

```
- include: pool_names.yml
  delegate_to: "{{ groups['ceph_monitors'][0] }}"
  when:
    - inventory_hostname == play_hosts[0]
    - ssd_pool is undefined
 ```